moved function to utils folder
